+++ core/src/pl/space_marine/game/impediments/enemies/Dragon.java
@@ -4,7 +4,7 @@ import pl.space_marine.game.assets.Image;
 import pl.space_marine.game.impediments.Enemy;
 
 public class Dragon extends Enemy {
-    public Dragon(Image image, float collisionDamage, float direction, int speed, int x, int y) {
+    public Dragon(Image image, float collisionDamage, int direction, int speed, int x, int y) {
         super(image, null, collisionDamage, direction, speed, x, y);
     }
 }
